Uso de UI mas integracion de modo editar y agregar

+++ app/Http/Controllers/EmpleadoController.php
@@ -50,7 +50,10 @@ class EmpleadoController extends Controller
         }
 
         Empleado::insert($datosEmpleados);
-        return response()->json($datosEmpleados);
+        //return response()->json($datosEmpleados);
+        return redirect('empleado')->with('mensaje', 'Empleado agregado con exito');
+
+
     }
 
     /**
@@ -113,7 +116,14 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
+        $empleado = Empleado::findOrFail($id);
+
+        if(Storage::delete('public/'.$empleado->Foto)){
+            
+            Empleado::destroy($id);
+        
+        }
+        
         return redirect('empleado');
